backend/requirements.txt
# =============================================================================
# TrustLens Backend - Python Dependencies
# =============================================================================
# Install with: pip install -r requirements.txt
# Python version: 3.9+ recommended
# =============================================================================

# -----------------------------------------------------------------------------
# Web Framework & API
# -----------------------------------------------------------------------------
fastapi>=0.100.0              # FastAPI web framework for REST API
uvicorn[standard]>=0.23.0     # ASGI server to run FastAPI (includes websockets, watchfiles)
pydantic>=2.0.0               # Data validation and settings management
python-dotenv>=1.0.0          # Load environment variables from .env files

# -----------------------------------------------------------------------------
# Machine Learning & Deep Learning
# -----------------------------------------------------------------------------
tensorflow>=2.16.2,<2.17      # Deep learning framework for MLP and Autoencoder models
# tensorflow-macos==2.15.0    # Uncomment for Apple Silicon Macs
# tensorflow-metal==1.1.0     # Uncomment for Apple Silicon GPU acceleration

# -----------------------------------------------------------------------------
# Data Processing & Scientific Computing
# -----------------------------------------------------------------------------
numpy>=1.24.0                 # Numerical computing for array operations
scikit-learn>=1.3.0           # Machine learning utilities (scaler, calibrator)
joblib>=1.3.0                 # Serialization for sklearn objects (.joblib files)

# -----------------------------------------------------------------------------
# Explainable AI (XAI)
# -----------------------------------------------------------------------------
shap>=0.42.0                  # SHAP (SHapley Additive exPlanations) for model interpretability

# -----------------------------------------------------------------------------
# Database & Authentication
# -----------------------------------------------------------------------------
pymongo>=4.8.0                # MongoDB driver with async client support
google-auth>=2.30.0           # Google ID token verification

# -----------------------------------------------------------------------------
# Optional: Development & Testing
# -----------------------------------------------------------------------------
# pytest>=7.0.0               # Testing framework
# httpx>=0.24.0               # Async HTTP client for testing FastAPI
# python-multipart>=0.0.6     # For file uploads (if needed)
